+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/EditActivity.kt
@@ -27,11 +27,12 @@ import com.simplemobiletools.commons.helpers.REAL_FILE_PATH
 import com.simplemobiletools.commons.models.FileDirItem
 import com.simplemobiletools.gallery.pro.R
 import com.simplemobiletools.gallery.pro.adapters.FiltersAdapter
+import com.simplemobiletools.gallery.pro.dialogs.OtherAspectRatioDialog
 import com.simplemobiletools.gallery.pro.dialogs.ResizeDialog
 import com.simplemobiletools.gallery.pro.dialogs.SaveAsDialog
 import com.simplemobiletools.gallery.pro.extensions.config
 import com.simplemobiletools.gallery.pro.extensions.openEditor
-import com.simplemobiletools.gallery.pro.helpers.FilterThumbnailsManager
+import com.simplemobiletools.gallery.pro.helpers.*
 import com.simplemobiletools.gallery.pro.models.FilterItem
 import com.theartofdev.edmodo.cropper.CropImageView
 import com.zomato.photofilters.FilterPack
@@ -54,11 +55,6 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
     private val ASPECT_Y = "aspectY"
     private val CROP = "crop"
 
-    private val ASPECT_RATIO_FREE = 0
-    private val ASPECT_RATIO_ONE_ONE = 1
-    private val ASPECT_RATIO_FOUR_THREE = 2
-    private val ASPECT_RATIO_SIXTEEN_NINE = 3
-
     // constants for bottom primary action groups
     private val PRIMARY_ACTION_NONE = 0
     private val PRIMARY_ACTION_FILTER = 1
@@ -71,6 +67,7 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
     private lateinit var saveUri: Uri
     private var resizeWidth = 0
     private var resizeHeight = 0
+    private var lastOtherAspectRatio: Pair<Int, Int>? = null
     private var currPrimaryAction = PRIMARY_ACTION_NONE
     private var currCropRotateAction = CROP_ROTATE_NONE
     private var currAspectRatio = ASPECT_RATIO_FREE
@@ -133,6 +130,11 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
 
         loadDefaultImageView()
         setupBottomActions()
+
+        if (config.lastEditorCropAspectRatio == ASPECT_RATIO_OTHER) {
+            lastOtherAspectRatio = Pair(config.lastEditorCropOtherAspectRatioX, config.lastEditorCropOtherAspectRatioY)
+        }
+        updateAspectRatio(config.lastEditorCropAspectRatio)
     }
 
     override fun onResume() {
@@ -328,6 +330,16 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
         bottom_aspect_ratio_sixteen_nine.setOnClickListener {
             updateAspectRatio(ASPECT_RATIO_SIXTEEN_NINE)
         }
+
+        bottom_aspect_ratio_other.setOnClickListener {
+            OtherAspectRatioDialog(this, lastOtherAspectRatio) {
+                lastOtherAspectRatio = it
+                config.lastEditorCropOtherAspectRatioX = it.first
+                config.lastEditorCropOtherAspectRatioY = it.second
+                updateAspectRatio(ASPECT_RATIO_OTHER)
+            }
+        }
+
         updateAspectRatioButtons()
     }
 
@@ -400,6 +412,7 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
 
     private fun updateAspectRatio(aspectRatio: Int) {
         currAspectRatio = aspectRatio
+        config.lastEditorCropAspectRatio = aspectRatio
         updateAspectRatioButtons()
 
         crop_image_view.apply {
@@ -409,7 +422,8 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
                 val newAspectRatio = when (aspectRatio) {
                     ASPECT_RATIO_ONE_ONE -> Pair(1, 1)
                     ASPECT_RATIO_FOUR_THREE -> Pair(4, 3)
-                    else -> Pair(16, 9)
+                    ASPECT_RATIO_SIXTEEN_NINE -> Pair(16, 9)
+                    else -> Pair(lastOtherAspectRatio!!.first, lastOtherAspectRatio!!.second)
                 }
 
                 setAspectRatio(newAspectRatio.first, newAspectRatio.second)
@@ -418,7 +432,7 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
     }
 
     private fun updateAspectRatioButtons() {
-        arrayOf(bottom_aspect_ratio_free, bottom_aspect_ratio_one_one, bottom_aspect_ratio_four_three, bottom_aspect_ratio_sixteen_nine).forEach {
+        arrayOf(bottom_aspect_ratio_free, bottom_aspect_ratio_one_one, bottom_aspect_ratio_four_three, bottom_aspect_ratio_sixteen_nine, bottom_aspect_ratio_other).forEach {
             it.setTextColor(Color.WHITE)
         }
 
@@ -426,7 +440,8 @@ class EditActivity : SimpleActivity(), CropImageView.OnCropImageCompleteListener
             ASPECT_RATIO_FREE -> bottom_aspect_ratio_free
             ASPECT_RATIO_ONE_ONE -> bottom_aspect_ratio_one_one
             ASPECT_RATIO_FOUR_THREE -> bottom_aspect_ratio_four_three
-            else -> bottom_aspect_ratio_sixteen_nine
+            ASPECT_RATIO_SIXTEEN_NINE -> bottom_aspect_ratio_sixteen_nine
+            else -> bottom_aspect_ratio_other
         }
 
         currentAspectRatioButton.setTextColor(config.primaryColor)