+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/SettingsActivity.kt
@@ -57,6 +57,7 @@ class SettingsActivity : SimpleActivity() {
         setupAllowVideoGestures()
         setupAllowDownGesture()
         setupBottomActions()
+        setupThumbnailVideoDuration()
         setupShowMediaCount()
         setupKeepLastModified()
         setupShowInfoBubble()
@@ -193,6 +194,14 @@ class SettingsActivity : SimpleActivity() {
         settings_crop_thumbnails_holder.setOnClickListener {
             settings_crop_thumbnails.toggle()
             config.cropThumbnails = settings_crop_thumbnails.isChecked
+        }
+    }
+
+    private fun setupThumbnailVideoDuration() {
+        settings_show_thumbnail_video_duration.isChecked = config.showThumbnailVideoDuration
+        settings_show_thumbnail_video_duration_holder.setOnClickListener {
+            settings_show_thumbnail_video_duration.toggle()
+            config.showThumbnailVideoDuration = settings_show_thumbnail_video_duration.isChecked
         }
     }