+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/receivers/RefreshMediaReceiver.kt
@@ -16,7 +16,7 @@ class RefreshMediaReceiver : BroadcastReceiver() {
 
         Thread {
             val medium = Medium(null, path.getFilenameFromPath(), path, path.getParentPath(), System.currentTimeMillis(), System.currentTimeMillis(),
-                    File(path).length(), getFileType(path), false, 0L)
+                    File(path).length(), getFileType(path), 0, false, 0L)
             context.galleryDB.MediumDao().insert(medium)
         }.start()
     }